SCRUM-63 move view models and handlers to domain from domain.model
