+++ app/src/main/java/com/air_wheelly/wheelly/domain/CarListHandler.kt
@@ -1,4 +1,4 @@
-package com.air_wheelly.wheelly.domain.model
+package com.air_wheelly.wheelly.domain
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
