+++ app/src/main/java/com/air_wheelly/wheelly/domain/CarViewModel.kt
@@ -1,3 +1,5 @@
+package com.air_wheelly.wheelly.domain
+
 import android.Manifest
 import android.annotation.SuppressLint
 import android.content.Context
